Extract duped Profile logic from My and Public
