+++ src/components/Profile/components/Identification/index.scss
@@ -1,4 +1,4 @@
-@import '../../../vars';
+@import '../../../../vars';
 
 $image-border-radius: 50%;
 // Main Image
