+++ src/components/Profile/components/MembershipsList/components/MembershipInfoCard/index.scss
@@ -1,4 +1,4 @@
-@import '../../../../../vars';
+@import '../../../../../../vars';
 
 .membership-info-card {
   &-description {
