+++ src/components/Profile/components/MembershipsList/index.scss
@@ -1,4 +1,4 @@
-@import '../../../vars';
+@import '../../../../vars';
 
 .memberships {
   &-header {
