+++ src/components/Profile/components/OfficeInfoList/index.scss
@@ -1,4 +1,4 @@
-@import '../../../vars';
+@import '../../../../vars';
 
 .office-info-list {
   background-color: $alternative-red;
