+++ src/components/Profile/components/PersonalInfoList/index.scss
@@ -1,4 +1,4 @@
-@import '../../../vars';
+@import '../../../../vars';
 
 .personal-info-list {
   &-header {
