+++ src/components/Profile/components/ProfileInfoListItem/index.scss
@@ -1,4 +1,4 @@
-@import '../../../vars.scss';
+@import '../../../../vars.scss';
 
 .profile-info-list-item {
   hyphens: auto;
