+++ src/components/Profile/components/SchedulePanel/components/RemoveScheduleDialog/index.js
