+++ src/components/Profile/components/VictoryPromiseDisplay/VictoryPromise.scss
