+++ src/components/Profile/components/index.js
@@ -1,0 +1,7 @@
+export { default as Identification } from './Identification';
+export { default as MembershipsList } from './MembershipsList';
+export { default as OfficeInfoList } from './OfficeInfoList';
+export { default as PersonalInfoList } from './PersonalInfoList';
+export { default as ProfileInfoListItem } from './ProfileInfoListItem';
+export { default as SchedulePanel } from './SchedulePanel';
+export { default as VictoryPromiseDisplay } from './VictoryPromiseDisplay';
